Remove Collector package, and put everything in rpc/core package to avoid cyclic imports.
